refactor: move post creation to utils
