+++ src/constants.ts
@@ -1,4 +1,5 @@
 export const channelId = '@reatex_channel';
+export const baseUrl = 'https://2ch.hk';
 export const url = 'https://2ch.hk/b/catalog_num.json';
 export const corsProxyHost = '0.0.0.0';
 export const corsProxyPort = 8080;
